wersja 0.91 - dajDzialy dalej do poprawy

+++ src/kadry/pracownik.java
@@ -43,7 +43,7 @@ class Pracownik{
 
     }
 
-    public double dzial(){
+    public int dzial(){
         return this.dzial;
 
     }
